Edited an error with the segue code
I was trying to remember from memory how to program the segue from code.
Instead of creating a function to prepare for the segue, i was calling
the segue twice in the code. Renamed the second function to prepare and
added a variable to store the url in the new Play Sounds VC.